Address Greg's feedback
