+++ lib/react_dom_server.dart
@@ -23,13 +23,6 @@ var renderToStaticMarkup;
 /// Sets configuration based on passed functions.
 ///
 /// Passes arguments to global variables.
-///
-/// > __DEPRECATED.__
-/// >
-/// > This library no longer supports `ReactDomServer` and static markup rendering.
-/// > This will be removed in 6.0.0, along with other configuration setting functions.
-@Deprecated(
-    'This library no longer supports `ReactDomServer` and static markup rendering. Will be removed from this library in the 6.0.0 release.')
 setReactDOMServerConfiguration(customRenderToString, customRenderToStaticMarkup) {
   renderToString = customRenderToString;
   renderToStaticMarkup = customRenderToStaticMarkup;
