+++ test/factory/common_factory_tests.dart
@@ -5,14 +5,14 @@ import 'dart:js';
 import 'dart:js_util';
 
 import 'package:meta/meta.dart';
+import 'package:react/react_client/component_factory.dart';
 import 'package:react/react_client/js_backed_map.dart';
-import 'package:react/react_client/utils.dart';
 import 'package:test/test.dart';
 
+import 'package:react/react.dart' as react;
 import 'package:react/react_client.dart';
 import 'package:react/react_dom.dart' as react_dom;
 import 'package:react/react_test_utils.dart' as rtu;
-import 'package:react/react.dart' as react;
 import 'package:react/react_client/react_interop.dart';
 
 import '../util.dart';
