+++ test/react_client_test.dart
@@ -11,9 +11,9 @@ import 'package:test/test.dart';
 
 import 'package:react/react.dart' as react;
 import 'package:react/react_client.dart';
+import 'package:react/react_client/component_factory.dart';
 import 'package:react/react_client/react_interop.dart' show React, ReactComponent;
 import 'package:react/react_client/js_interop_helpers.dart';
-import 'package:react/react_client/utils.dart';
 import 'package:react/react_dom.dart' as react_dom;
 import 'package:react/src/react_client/event_prop_key_to_event_factory.dart';
